+++ dojo/tools/openvas/parser.py
@@ -1,5 +1,7 @@
-from dojo.tools.openvas.csv_parser import OpenVASCSVParser
-from dojo.tools.openvas.xml_parser import OpenVASXMLParser
+from dojo.tools.openvas.parser_v1.csv_parser import OpenVASCSVParser
+from dojo.tools.openvas.parser_v1.xml_parser import OpenVASXMLParser
+from dojo.tools.openvas.parser_v2.csv_parser import OpenVASCSVParserV2
+from dojo.tools.openvas.parser_v2.xml_parser import OpenVASXMLParserV2
 
 
 class OpenVASParser:
@@ -17,4 +19,22 @@ class OpenVASParser:
             return OpenVASCSVParser().get_findings(filename, test)
         if str(filename.name).endswith(".xml"):
             return OpenVASXMLParser().get_findings(filename, test)
+        return None
+
+
+class OpenVASParserV2:
+    def get_scan_types(self):
+        return ["OpenVAS Parser v2"]
+
+    def get_label_for_scan_types(self, scan_type):
+        return scan_type
+
+    def get_description_for_scan_types(self, scan_type):
+        return "Import CSV or XML output of Greenbone OpenVAS report."
+
+    def get_findings(self, filename, test):
+        if str(filename.name).endswith(".csv"):
+            return OpenVASCSVParserV2().get_findings(filename, test)
+        if str(filename.name).endswith(".xml"):
+            return OpenVASXMLParserV2().get_findings(filename, test)
         return None
