+++ dojo/tools/openvas/parser_v1/xml_parser.py
