+++ dojo/tools/openvas/parser_v2/xml_parser.py
@@ -4,7 +4,7 @@ from xml.dom import NamespaceErr
 from defusedxml import ElementTree
 
 from dojo.models import Endpoint, Finding
-from dojo.tools.openvas_v2.common import (
+from dojo.tools.openvas.parser_v2.common import (
     OpenVASFindingAuxData,
     cleanup_openvas_text,
     deduplicate,
