--- dojo/tools/openvas_v2/__init__.py
@@ -1,1 +1,0 @@
-__author__ = "manuel-sommer"
